services/orchestrator/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
pyjwt[crypto]==2.8.0
weaviate-client==3.25.0
redis==5.0.0
pytest==7.4.0
pytest-asyncio==0.21.0